Correct the ordering of metrics
When common metrics were moved from a yml file to the DB, the priority
values were not added to the DB.
In this commit, a priority column is added to the prometheus_metrics
table and common metric groups are ordered in descending order of
priority. All metrics in a group are expected to have the same priority.

Custom metrics do not have a priority and will default to priority 0,
as before.