Initialize FactoryBot Commits with Commit#from_hash
This change initializes a Commit with its @raw variable as a
Gitlab::Git::Commit rather than an OpenStruct as it was previously.

Commit#method_missing will now behave as expected, and call the method
on an instance of Gitlab::Git::Commit, allowing us to use (for example)
Commit#safe_message.

Added skip_create, as the Commit class is ORM-less and is not persisted.
This allows a user to use the create strategy with this factory and
FactoryBot won't call save!.