Avoid accidental additional calls to SquashService
MergeRequests::MergeService previously would call
MergeRequests::SquashService#execute more than once during a squash, as
the `source` method within that class was called multiple times.

Using memoization of the squash_sha so that SquashService#execute is
only ever called once, and the result of the first call can be returned.