Don't configure clusters immediately after they are created
Immediate configuration is not ideal for group and instance
level clusters as projects that may never be deployed would
still have Kubernetes namespaces and service accounts created
for them.

The next step will be to create only the resources that
are required for the project being deployed, at the time
of deployment.