Fix Auto DevOps missing domain error handling
For some reason this value can be an empty string and we want to ensure that we
correctly fail the job before a deploy. This is particularly problematic at the
moment when coupled with https://gitlab.com/gitlab-org/gitlab-ce/issues/54760
as the user has no easy way to recover without destroying the cluster and
starting over.