Confirm existence of head_pipeline with retries
Pipelines are created by an async worker, so a rapid sequence of API
calls can trigger a state where the pipeline, whose existence is part of
determining if we wait for the pipeline to successfully complete before
merging, can trigger the MR to be immediately merged instead. By
introducing a slight delay, we give our async worker a chance to
process.

Closes #55127