Remove extraneous #banzai_render_context
Most/all of the work we're doing in this method is done at creation/edit
time, so do we need to also check at render? Let's find out! Committing
this for a WIP branch in order to get a full test run to see what the
automated testing thinks might be broken by removing this.