Retry fetching Kubernetes Secret#token (#63507)
Since Kubernetes is creating the Secret and token asynchronously it is
necessary that we implement some delay or retrying logic to avoid a race
condition where we fetch a Secret before the token is even set. There does not
appear to be any way for us to force it to be set with any synchronous API call
so retrying seems to be the only option.