Fix whitespace in ci and dev docs
Many code blocks are 4spaced, and they render in GitLab
without coloring as a result, even though they are
fenced with a language label. If in a list, other items
will render as being in a code block too, even if not
meant to. This fixes all these issues for most ci and dev
docs, and cleans up minor whitespace issues as well.