Bring backward compatibility for request profiles
It seems that we missed the backward compatibility support for profiles
in the existing folder.

This commit also fixes some specs to be idempotent and work in a temporary directory
which not always seems to be the case.

This commit also brings the profile_spec.rb which seems to be missing.