Invalidate branches cache on MR refresh
Whenever `UpdateMergeRequestsWorker` is enqueued and
`MergeRequests::RefreshService` is called, it'll check if the
source branch of each MR asssociated to the push exists or not
via `MergeRequest#source_branch_exists?`. The said method will
call `Repository#branch_exists?` which is cached in `Rails.cache`.

When the cache contains outdated data and the source branch actually
exists, the `MergeRequests#RefreshService` job will close associated MRs
which is not correct.

The fix is to expire the branches cache of the source project so we
can get updated information when determining if we need to close the
MR or not.