Fix style-lint errors in wiki.scss
There were two errors: duplicate styles and @extend used with a class
name. I added `.top-area` class to all the instances where
`.wiki-page-header` is used since %top-area didn't really make a lot of
sense as a placeholder selector. This is also consistent with behaviour
on other pages where the `.top-area` class is used alongside the other
classes for the specific header styling.