Add #diff_stats route to fetch basic path/stats data
This is a very basic implementation of an endpoint to expose a
Gitlab::Git::DiffStatsCollection of path and +/- stats for a given MR.
It assumes you want the diff stats for the current HEAD of the MR versus
the target branch. The response includes an assembled `diff_for_path`
URL which should return the highlighted diff for the corresponding path
or file.

Things we are very much not doing here:
+ any kind of explicit caching
+ feature flagging
+ metrics gathering
+ fancy logging

This is a Proof of Concept implementation, and should as least be
feature flagged before deployment.