Set raw content when re-naming files
This allows the editor to show correct content when showing renamed
files on the staged/unstaged views. Otherwise the original content of
the renamed file is kept.

Also, even though we rename a file, when the file is also changed we
explicitly send `content` to API for proper entry's creation