Set default authorized_keys_file
This is the same as gitlab-shell's default. This is to ensure
that it's always set.

It needs to be the same as gitlab-shell's default because we
don't set a default value in omnibus-gitlab. If users don't
set the value of that config in their install and they upgraded,
we must ensure that it's still going to point to the same
authorized keys file.